fix implicit numeric widening warnings

+++ axle-core/src/main/scala/axle/visualize/BarChart.scala
@@ -45,9 +45,9 @@ case class BarChart[C, Y, D, H](
         val dataView: DataView[C, Y, D],
         val lengthSpaceY: LengthSpace[Y, _, Double]) {
 
-  val titleText = title.map(Text(_, width / 2, titleFontSize, titleFontName, titleFontSize, bold = true))
-  val xAxisLabelText = xAxisLabel.map(Text(_, width / 2, height - border / 2, normalFontName, normalFontSize, bold = true))
-  val yAxisLabelText = yAxisLabel.map(Text(_, 20, height / 2, normalFontName, normalFontSize, bold = true, angle = Some(90d *: angleDouble.degree)))
+  val titleText = title.map(Text(_, width / 2d, titleFontSize.toDouble, titleFontName, titleFontSize.toDouble, bold = true))
+  val xAxisLabelText = xAxisLabel.map(Text(_, width / 2d, height - border / 2d, normalFontName, normalFontSize.toDouble, bold = true))
+  val yAxisLabelText = yAxisLabel.map(Text(_, 20, height / 2d, normalFontName, normalFontSize.toDouble, bold = true, angle = Some(90d *: angleDouble.degree)))
 
   val keyOpt = if (drawKey) {
     Some(BarChartKey(this, keyTitle))
